Removes unused code and adds confirm statement to delete fetch

+++ library-frontend/src/adapters/genreAdapter.js
@@ -2,16 +2,6 @@ class GenreAdapter {
 
   static url = "http://127.0.0.1:3000/genres"
 
-  // constructor() {
-  //   this.url = "http://127.0.0.1:3000/genres"
-  //   this.getForm()
-  // }
-
-  // getForm() {
-  //   const genreForm = document.getElementById("genre-form")
-  //   genreForm.addEventListener("submit", this.newGenre.bind(this))
-  // }
-
   static fetchGenres() {
     fetch(this.url)
     .then(res => res.json())
@@ -26,7 +16,6 @@ class GenreAdapter {
   static newGenre() {
     event.preventDefault()
 
-    // debugger
     fetch(GenreAdapter.url, {
       method: "POST",
       headers: {
@@ -48,15 +37,18 @@ class GenreAdapter {
 
   static deleteGenre() {
     event.preventDefault()
-    
-    fetch(`${GenreAdapter.url}/${this.id}`, {
-      method: "DELETE",
-      headers: {
-          "Content-Type": "application/json"
-        }
-    })
-    
-    document.getElementById(this.id).remove()
+    if ( confirm(`This will delete the entire` + 
+      ` ${this.name} genre and all the books in` +
+      ` it. Click OK if you want to proceed.`)
+      ) {
+      fetch(`${GenreAdapter.url}/${this.id}`, {
+        method: "DELETE",
+        headers: {
+            "Content-Type": "application/json"
+          }
+      })
+        
+      document.getElementById(this.id).remove()
+    }
   }  
-
 }
